Adding sanity checking to the EEPROM values
Calibration values read out from the EEPROM should be sanity checked to
make sure they are within semi-reasonable bounds. The offset (intercept)
value may deviate a couple of degrees from the factory, so we set a wide
range of +/- 3. The slope range is restricted to 0.5 to 1.5 by this
code, and we really ought to expect a slope very close to 1.0. If
calibrations values are outside these bounds, either the EEPROM values
are borked, or the calibration coefficients weren't properly determined.